Change how configuration files are managed.
The module in now based on a main configuration file and a directory
containing configuration files for commands. Configuration files are
recursively loaded but ignoring files and directories of others
commands. YAML and JSON files are loaded while other files are put as
is.

Main configuration file and commands' directory are retrieved from
`conf_file` and `conf_dir` arguments but have default values if these
options have no value (*conf.yml* file and *conf/* directory in the
main program directory).